finish services, add controller and error handling

+++ backend/EnterpriseTaskManagementApi/Services/TasksService.cs
@@ -1,4 +1,5 @@
 using EnterpriseTaskManagementApi.Context;
+using EnterpriseTaskManagementApi.CustomExceptions;
 using EnterpriseTaskManagementApi.Models;
 using Microsoft.EntityFrameworkCore;
 
@@ -17,6 +18,11 @@ public class TasksService : ITasksService
     {
         var entities = await _context.Tasks.ToListAsync();
 
+        if (entities is null)
+        {
+            throw new NotFoundException("No tasks found.");
+        }
+
         return entities;
     }
 
@@ -24,33 +30,63 @@ public class TasksService : ITasksService
     {
         var entity = await _context.Tasks.FindAsync(id);
 
-        //maybe it will return null, the verification is on the controller
-        return entity!;
+        if (entity is null)
+        {
+            throw new NotFoundException("Task with id: " + id + "was not found.");
+        }
+
+        return entity;
     }
 
-    public Task<IEnumerable<TaskItem>> AddTaskAsync()
+    public async Task<TaskItem> AddTaskServiceAsync(TaskItemDTO taskItem)
     {
-        throw new NotImplementedException();
+        var entity = new TaskItem(
+            taskItem.Title,
+            taskItem.Project,
+            taskItem.Responsible,
+            taskItem.Deadline,
+            taskItem.Status
+        );
+
+        await _context.Tasks.AddAsync(entity);
+        await _context.SaveChangesAsync();
+
+        return entity;
     }
 
-    public Task<TaskItem> UpdateTaskServiceAsync()
+    public async Task<TaskItem> UpdateTaskServiceAsync(Guid id, TaskItemDTO taskItemDTO)
     {
-        throw new NotImplementedException();
+        var entity = await _context.Tasks.FindAsync(id);
+
+        if (entity is null)
+        {
+            throw new NotFoundException("Task with id: " + id + "was not found.");
+        }
+
+        entity.Title = taskItemDTO.Title;
+        entity.Project = taskItemDTO.Project;
+        entity.Responsible = taskItemDTO.Responsible;
+        entity.Deadline = taskItemDTO.Deadline;
+        entity.Status = taskItemDTO.Status;
+
+        await _context.SaveChangesAsync();
+
+        return entity;
     }
 
     public async Task<TaskItem> DeleteTaskServiceAsync(Guid id)
     {
         var entity = await _context.Tasks.FindAsync(id);
 
+        if (entity is null)
+        {
+            throw new NotFoundException("Task with id: " + id + "was not found.");
+        }
+
         _context.Tasks.Remove(entity);
         await _context.SaveChangesAsync();
 
         return entity;
-    }
-
-    public Task<TaskItem> GetTaskByIdServiceAsync()
-    {
-        throw new NotImplementedException();
     }
 
 }
